+++ automotive/src/main/java/de/moleman1024/audiowagon/medialibrary/contenthierarchy/ContentHierarchyRootFiles.kt
@@ -120,8 +120,9 @@ class ContentHierarchyRootFiles(
         }
         val rootDirectory = Directory(storageLocation.getRootURI())
         val directoryContents = storageLocation.getDirectoryContentsPlayable(rootDirectory)
-        return directoryContents.filter { !it.name.matches(Util.DIRECTORIES_TO_IGNORE_REGEX) }
-            .sortedBy { it.name.lowercase() }
+        return directoryContents.filter {
+            !it.name.matches(Util.DIRECTORIES_TO_IGNORE_REGEX) && !it.name.matches(Util.FILES_TO_IGNORE_REGEX)
+        }.sortedBy { it.name.lowercase() }
     }
 
 }